+++ lerna.json
@@ -1,5 +1,6 @@
 {
   "packages": ["packages/*"],
   "version": "6.0.8",
-  "npmClient": "yarn"
+  "npmClient": "yarn",
+  "$schema": "node_modules/lerna/schemas/lerna-schema.json"
 }
