+++ package.json
@@ -4,10 +4,12 @@
   "license": "MIT",
   "description": "Declarative and type-safe model & validation library",
   "author": "Smartly.io",
+  "private": true,
   "repository": {
     "type": "git",
     "url": "https://github.com/smartlyio/yavl.git"
   },
+  "workspaces": ["packages/*"],
   "scripts": {
     "build": "lerna run --stream build",
     "clean": "lerna run clean",
@@ -23,9 +25,6 @@
     "ci:lint": "yarn lint",
     "ci:test": "yarn test"
   },
-  "peerDependencies": {
-    "typescript": ">=4 <6"
-  },
   "dependencies": {
     "ramda": "^0.28.0"
   },
@@ -39,11 +38,8 @@
     "validate"
   ],
   "devDependencies": {
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/react-hooks": "^8.0.1",
     "@types/jest": "29.5.2",
     "@types/ramda": "0.28.11",
-    "@types/react": "^18.2.28",
     "@typescript-eslint/eslint-plugin": "^5.60.0",
     "@typescript-eslint/parser": "^5.60.0",
     "eslint": "7.19.0",
@@ -52,12 +48,8 @@
     "eslint-plugin-prettier": "3.3.1",
     "eslint-plugin-react-hooks": "4.3.0",
     "jest": "29.5.0",
-    "jest-environment-jsdom": "^29.7.0",
     "prettier": "2.2.1",
     "prettier-eslint": "12.0.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-test-renderer": "^18.2.0",
     "ts-jest": "29.1.0",
     "typescript": "^5.1.3",
     "lerna": "^8.1.8"
