+++ packages/yavl-hooks/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@smartlyio/yavl",
+  "name": "@smartlyio/yavl-hooks",
   "version": "6.0.8",
   "license": "MIT",
-  "description": "Declarative and type-safe model & validation library",
+  "description": "Collection of React hooks for YAVL",
   "author": "Smartly.io",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -13,10 +13,14 @@
   "repository": {
     "type": "git",
     "url": "https://github.com/smartlyio/yavl.git",
-    "directory": "packages/yavl"
+    "directory": "packages/yavl-hooks"
   },
   "peerDependencies": {
     "typescript": ">=4 <6"
+  },
+  "dependencies": {
+    "@smartlyio/yavl": "^6.0.8",
+    "tsconfig-paths": "4.2.0"
   },
   "keywords": [
     "typescript",
@@ -27,6 +31,16 @@
     "validator",
     "validate"
   ],
+  "devDependencies": {
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/react-hooks": "^8.0.1",
+    "@types/react": "^18.2.28",
+    "jest-environment-jsdom": "^29.7.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-test-renderer": "^18.2.0",
+    "typescript": "^5.1.3"
+  },
   "files": [
     "dist/*"
   ],
