+++ packages/yavl-hooks/src/hooks/useFieldsAnnotation.ts
@@ -1,7 +1,5 @@
 import * as R from 'ramda';
-import { getAnnotationValue } from '../getAnnotationValue';
-import { Annotation } from '../types';
-import { ModelValidationContext } from '../validate/types';
+import { ModelValidationContext, Annotation, getAnnotationValue } from '@smartlyio/yavl';
 import { useAnnotations } from './useAnnotations';
 import { useMemo } from 'react';
 import { useMemoizedValue } from './useMemoizedValue';
