+++ packages/yavl-hooks/tsconfig.dist.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.dist.json",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "noEmit": false,
+    "declaration": true
+  },
+  "files": ["./src/index.ts"],
+  "include": []
+}
