+++ tsconfig.dist.json
@@ -2,7 +2,8 @@
   "extends": "./tsconfig.json",
   "compilerOptions": {
     "noEmit": false,
-    "declaration": true
+    "declaration": true,
+    "baseUrl": "./src"
   },
   "files": [],
   "include": []
